backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
slowapi==0.1.9
python-whois==0.7.3
dnspython==2.6.1
requests==2.32.3
httpx==0.27.0
aiosqlite==0.20.0
pydantic==2.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
lightgbm==4.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
jinja2==3.1.4
